osa 4 valmis

+++ osa4/blogilista/controllers/blogs.js
@@ -1,29 +1,61 @@
 const blogsRouter = require("express").Router();
 const Blog = require("../models/blog");
+const middleware = require("../utils/middleware");
 
-blogsRouter.get("/api/blogs", (request, response) => {
-  Blog.find({}).then((blogs) => {
-    response.json(blogs);
+blogsRouter.get("/", async (request, response) => {
+  const blogs = await Blog.find({}).populate("user", { username: 1, name: 1 });
+  response.json(blogs);
+});
+
+blogsRouter.get("/:id", async (request, response) => {
+  const blog = await Blog.findById(request.params.id);
+  if (blog) {
+    response.json(blog);
+  } else {
+    response.status(404).end();
+  }
+});
+
+blogsRouter.post("/", middleware.userExtractor, async (request, response) => {
+  const body = request.body;
+  const user = request.user;
+
+  if (body.title === undefined || body.url === undefined) {
+    return response.status(400).end();
+  }
+
+  const blog = new Blog({
+    title: body.title,
+    author: body.author,
+    url: body.url,
+    likes: body.likes || 0,
+    user: user.id,
   });
+  const savedBlog = await blog.save();
+  user.blogs = user.blogs.concat(savedBlog._id);
+  await user.save();
+
+  response.json(savedBlog);
 });
 
-blogsRouter.post("/api/blogs", (request, response) => {
-  const blog = new Blog(request.body);
+blogsRouter.delete(
+  "/:id",
+  middleware.userExtractor,
+  async (request, response) => {
+    const user = request.user;
 
-  blog
-    .save()
-    .then((result) => {
-      response.status(201).json(result);
-    })
-    .catch((e) => response.status(400).json(e));
-});
+    const blog = await Blog.findById(request.params.id);
 
-blogsRouter.delete("/api/blogs/:id", async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id);
-  response.status(204).end();
-});
+    if (blog.user.toString() !== user._id.toString()) {
+      return response.status(401).json({ error: "invalid token" });
+    }
 
-blogsRouter.put("/api/blogs/:id", (request, response, next) => {
+    await Blog.findByIdAndRemove(request.params.id);
+    response.status(204).end();
+  }
+);
+
+blogsRouter.put("/:id", (request, response, next) => {
   const body = request.body;
 
   const blog = {
